Added RBIT function, minor updates.
Added reverseBits32 function.
Fixed type of the size field in class descriptors.
Removed redundant includes.

+++ include/entity.h
@@ -7,10 +7,11 @@
 #ifndef XCORE_ENTITY_H_
 #define XCORE_ENTITY_H_
 /*----------------------------------------------------------------------------*/
+#include <stddef.h>
 #include <error.h>
 /*----------------------------------------------------------------------------*/
 #define CLASS_HEADER \
-    unsigned int size;\
+    size_t size;\
     enum result (*init)(void *, const void *);\
     void (*deinit)(void *);
